Fix disabling of `'list'` in `man` filetype
Prior to this commit, the ftplugin/man.vim file in my .vim directory was
run instead of the ftplugin/man.vim containing the actual Man.vim
plugin, resulting in the `:Man` command being unavailable and the
`'list'` setting to be disabled by default on all buffers instead of
only being disabled in `man` buffers.

These bugs are fixed by moving the disabling of the `list` setting into
my vimrc, putting it into an `:autocmd` to be run on `man` buffers.