Fix terminal-mode mapping for `:tabclose`
I made a mapping for closing a tab (`<leader>x`) in normal mode, and I
made a terminal-mode equivalent (`<C-W>x`) for doing the same, without
needing to press the leader key after pressing the `'termwinkey'` for
convenience.

Unfortunately this mapping overrode an existing command for switching
the positions of windows within a tab; this is fixed by pressing the
leader key, just like in normal mode.