Set up JDTLS
This was difficult to implement, especially due to the fact that I'm
using Git Bash on Windows, which meant that I had to do extra work while
trying to make the code as portable as possible.

I made the choice to make the LSP feature opt-in rather than having it
launch on every single Java file, since I think that launching a server
to make a quick edit is sort of silly.