Don't ask to install vim-plug if user has already declined
We do so by using a Vim feature which saves uppercased global variables
between Vim sessions; see `:h viminfo-!` for Vim or `:h shada-!` for
Neovim.  When the user says they don't want to install vim-plug, we
set the variable.  All we have to do then is to check whether the
variable exists, and if it does, don't run the install script again.