Move all Parinfer replacement `autocmd`s into one `augroup`
Prior to this commit, there was one global `augroup`, which provided an
autocmd which called a Vim script (`parinfer_augroup.vim`) when the
buffer's `filetype` matched `g:vim_parinfer_filetypes`, which made it so
that each buffer had its own `augroup` with the corresponding
`autocmd`s.  This is inconvenient, since disabling `parinfer` required
deleting the `augroup`s for each buffer plus the main `parinfer-setup`
`augroup`.

Now, there is only one `augroup` (which has been renamed to `parinfer`
to better express the fact that it is a replacement for the `augroup`
provided in the plugin), which sets up the `autocmd`s directly when the
buffer's `filetype` matches `g:vim_parinfer_filetypes`, thereby allowing
me to get rid of the `parinfer_augroup.vim` file and making it easier to
disable parinfer (I can just delete the `parinfer` `augroup`).  Plus,
it's cleaner.