Detect if `:h viminfo-!` is supported
This is used in the vim-plug install script.  If `:h viminfo-!` (or `:h
shada-!` in Neovim) is supported, the user can just run
    let g:NO_INSTALL_PLUG = v:true
they don't have to add that line into their vimrc; just run the script
and viminfo (shada) will keep that variable in memory.

Thus, in this commit, we give the user the relevant instruction in case
they choose not to install vim-plug.