Add mappings to interact with the system clipboard
About `<leader>p`: it makes more sense to me to make the mapping for
pasting from the system clipboard to `<leader>p` (since `p` is the
pasting command).  However, this conflicts with the previously-defined
`<leader>p` mapping which populates the buffer with the path to a
temporary filename, so I had to redefine that mapping to `<leader>t`.