Always use inner word for the `S` mapping
I created the `S` mapping for deleting `count` words and entering insert
mode.  I originally used `ciw` for prefixing a mapping, but this didn't
work well for a count since the "inner word" text object selects word
boundaries when the count is greater than two, meaning that this mapping
would actually delete `(count + 1) / 2` words (rounded down).

A bodge I had used until then was to use the "a word" text object when a
count was given.  Although this fulfills the requirement that the
mapping of deleting `count` words, it also deletes surrounding spaces,
meaning that I had to enter a space when I finish typing the replacement
for the text I has deleted.

I fixed this in the commit by using the "inner word" text object; when a
count is specified, we double it and take one from it to ensure that it
removes `count` words from the text (since
`((count * 2) - 1 + 1) / 2 = count`).

To avoid `:h motion-count-multiplied` from applying to our command, we
must prefix the `ciw` with `<c-\><c-n>` (not `<esc>` as that enters
insert mode when normal mode was entered from insert mode via `<c-o>`)
to remove the count Vim adds to the `ciw` command (otherwise, it would
remove `count^2 - count / 2` words, rounded down). Since `<c-\><c-n>`
discards the register used, we must provide it in the mapping, hence the
`v:register` part.