Support `:<count>[TS]Indent` syntax
Prior to this commit, the count for `:TI` or `:SI` (and the like) had to
be supplied after the command, preceded by a space.  For convenience –
and to be consistent with Vim's interface for other commands that take a
count, the command now allows the count to be given before their name by
using `-count` and `<count>` in their definition instead of `-nargs=?`
and `<args>`.